+++ v3/examples/plain/main.go
@@ -42,7 +42,7 @@ func main() {
 		JS:    `window.iamhere = function() { console.log("Hello World!"); }`,
 	})
 
-	app.customEventProcessor.On("clicked", func(_ *application.CustomEvent) {
+	app.OnEvent("clicked", func(_ *application.CustomEvent) {
 		println("clicked")
 	})