+++ v3/plugins/experimental/start_at_login/plugin.go
@@ -1,6 +1,7 @@
 package start_at_login
 
 import (
+	"context"
 	"github.com/wailsapp/wails/v3/pkg/application"
 	"io/fs"
 )
@@ -22,9 +23,9 @@ func NewPlugin(options Config) *Plugin {
 	}
 }
 
-// Shutdown is called when the app is shutting down
+// OnShutdown is called when the app is shutting down
 // You can use this to clean up any resources you have allocated
-func (p *Plugin) Shutdown() error { return nil }
+func (p *Plugin) OnShutdown() error { return nil }
 
 // Name returns the name of the plugin.
 // You should use the go module format e.g. github.com/myuser/myplugin
@@ -32,7 +33,7 @@ func (p *Plugin) Name() string {
 	return "github.com/wailsapp/wails/v3/plugins/start_at_login"
 }
 
-func (p *Plugin) Init(api application.PluginAPI) error {
+func (p *Plugin) OnStartup(ctx context.Context, options application.ServiceOptions) error {
 	// OS specific initialiser
 	err := p.init()
 	if err != nil {